+++ src/envoy/http/authnv2/authenticator_base.cc
@@ -13,11 +13,11 @@
  * limitations under the License.
  */
 
-#include "src/envoy/http/authn/authenticator_base.h"
+#include "src/envoy/http/authnv2/authenticator_base.h"
 
 #include "common/common/assert.h"
 #include "common/config/metadata.h"
-#include "src/envoy/http/authn/authn_utils.h"
+#include "src/envoy/http/authnv2/authn_utils.h"
 #include "src/envoy/utils/filter_names.h"
 #include "src/envoy/utils/utils.h"
