+++ src/envoy/http/authnv2/authenticator_base.h
@@ -17,7 +17,7 @@
 
 #include "authentication/v1alpha1/policy.pb.h"
 #include "common/common/logger.h"
-#include "src/envoy/http/authn/filter_context.h"
+#include "src/envoy/http/authnv2/filter_context.h"
 #include "src/istio/authn/context.pb.h"
 
 namespace Envoy {
