+++ src/envoy/http/authnv2/authenticator_base_test.cc
@@ -13,14 +13,14 @@
  * limitations under the License.
  */
 
-#include "src/envoy/http/authn/authenticator_base.h"
+#include "src/envoy/http/authnv2/authenticator_base.h"
 
 #include "common/common/base64.h"
 #include "common/protobuf/protobuf.h"
 #include "envoy/api/v2/core/base.pb.h"
 #include "envoy/config/filter/http/authn/v2alpha1/config.pb.h"
 #include "gmock/gmock.h"
-#include "src/envoy/http/authn/test_utils.h"
+#include "src/envoy/http/authnv2/test_utils.h"
 #include "src/envoy/utils/filter_names.h"
 #include "test/mocks/network/mocks.h"
 #include "test/mocks/ssl/mocks.h"
