+++ src/envoy/http/authnv2/filter_context.cc
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-#include "src/envoy/http/authn/filter_context.h"
+#include "src/envoy/http/authnv2/filter_context.h"
 
 #include "src/envoy/utils/filter_names.h"
 #include "src/envoy/utils/utils.h"
