+++ src/envoy/http/authnv2/http_filter.cc
@@ -13,13 +13,13 @@
  * limitations under the License.
  */
 
-#include "src/envoy/http/authn/http_filter.h"
+#include "src/envoy/http/authnv2/http_filter.h"
 
 #include "authentication/v1alpha1/policy.pb.h"
 #include "common/http/utility.h"
 #include "envoy/config/filter/http/authn/v2alpha1/config.pb.h"
-#include "src/envoy/http/authn/origin_authenticator.h"
-#include "src/envoy/http/authn/peer_authenticator.h"
+#include "src/envoy/http/authnv2/origin_authenticator.h"
+#include "src/envoy/http/authnv2/peer_authenticator.h"
 #include "src/envoy/utils/authn.h"
 #include "src/envoy/utils/filter_names.h"
 #include "src/envoy/utils/utils.h"
