+++ src/envoy/http/authnv2/http_filter_factory.cc
@@ -17,7 +17,7 @@
 #include "envoy/registry/registry.h"
 #include "envoy/server/filter_config.h"
 #include "google/protobuf/util/json_util.h"
-#include "src/envoy/http/authn/http_filter.h"
+#include "src/envoy/http/authnv2/http_filter.h"
 #include "src/envoy/utils/filter_names.h"
 #include "src/envoy/utils/utils.h"
