+++ src/envoy/http/authnv2/origin_authenticator.cc
@@ -13,12 +13,12 @@
  * limitations under the License.
  */
 
-#include "src/envoy/http/authn/origin_authenticator.h"
+#include "src/envoy/http/authnv2/origin_authenticator.h"
 
 #include "absl/strings/match.h"
 #include "authentication/v1alpha1/policy.pb.h"
 #include "common/http/headers.h"
-#include "src/envoy/http/authn/authn_utils.h"
+#include "src/envoy/http/authnv2/authn_utils.h"
 
 using istio::authn::Payload;
